Update lib/XmlFields.php
We found that if we have a lot of items (specifically, more than 10) the API throws back an error message:

Error validating xml data against the schema on line 2 tag name "lineItemData10" is not allowed. Possible tag names are: [lineItemData]

This seems to be the best fix--but it's a quick one. Maybe you guys can double-check to make sure there aren't similar problems elsewhere.